Test cleanup and feature addition
Implement https://github.com/elastic/logstash-contrib/pull/118 in
a way that makes sense to this plugin.

Add tests for the timestamp changes

Clean up the tests and make the let() statements do their job

Randomize the container names a bit to prevent collisions in the
event of a hung test.

Fixes #2